Dungeon Generation - Rooms "Finding" Each other
Rooms have the ability to set their exits now, In the dungeon generator, I need to spawn the Tiles currently setup. The dungeon will generate a list of rooms, starting with a Start room, then x number of Transition Tiles, until it gets to the # of Rooms to Spawn - 1, in which it will spawn a Boss room, after which it spawns a end room, to move to the next level (These rooms are not built yet, still working on generation)

When I have them Spawning, I want to position them (firstly) in a line, and have the Start to End room All connect with proper Exits. Exits will be determined when the tile is spawned, the next tile will spawn at the next exit position, with the "entrance" exit already setup, then it will determine its new exit point. When I have the " Line Dungeon" Setup, I will start working on Branching Dungeon paths.

After I have that all Setup, the Next thing is to add additional tiles to the Map, at Exits generated in the first step of generation, (Eg, Start room as a N exit, Next room has a S Exit (the entrance you came from) a N Exit (To the boss Room) and a E Exit, the generator will go to the boss room, then spawn the end room, after which, each Tile in the New Dungeon Array will be looked at, and if theres an exit not used already, we will spawn a new tile, New Exits only spawn if Max Additional Rooms is < Current Number of Rooms. Max Additional rooms will be the number to say how big the Dungeon is, # Of Rooms to Spawn is the Length of the Dungeon if you gro from A to B straight.)